SEO Tools: make feature available on non-connected sites. (#36528)
* SEO Tools: make feature available on non-connected sites.

The module was previously restricted to paid Jetpack plans. When we added the options to customize the different SEO options to wp-admin (in the block editor and in Jetpack settings), we made the module available to all plans. It is still restricted to paid plans on WordPress.com though.

As a result:

- On self-hosted, we do not need to be checking if the feature is available; it always is.
- On WordPress.com, we should be checking if the feature is available.
- On self-hosted, the feature can consequently be available on offline mode.

* Add missing parameter to stub function

The original function on WordPress.com simple does include that second parameter.

* Update Phan config

* Bump version

* Update plugin registration to match simplified module availability

* Bump version

* Add plan check back to editor extension

It does not currently make any calls to WordPress.com behind the scenes. `set_availability_for_plan` relies on `Current_Plan::supports()` behind the scenes, but without passing a `$refresh_from_wpcom` second parameter so it only relies on local data on self-hosted sites.

* Remove offline mode check in wp-admin UI

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8468223482